added payment using stripe and completed the wallet controllers and entity

+++ src/main/java/com/nrifintech/cms/controllers/OrderController.java
@@ -201,10 +201,10 @@ public class OrderController {
 
 					order.setCartItems(new ArrayList<>(cartItems));
 					
-					Double orderTotal=0.0;
+					Integer orderTotal=0;
 					
 					for(CartItem item : order.getCartItems()) {
-						Double price = item.getPrice() * item.getQuantity();
+						Integer price = (int) (item.getPrice() * item.getQuantity());
 						orderTotal += price;	
 					};
